move belle graph utils into separate module
